replica_pinning: add C++ unit and deterministic tests
Add C++ test coverage for replica pinning:

Planner tests: repair moves replicas to higher-priority racks,
violation detection, no-op when optimal, multiple violations, and
fill-then-overflow tests for rack awareness ON/OFF with varying
capacity scenarios.

Scoring tests (separate file): constraint scoring verification
(group 0 = max_score, group k = max_score/(k+1), same group equality,
unpreferred = 0, rackless = 0).

Simulator tests (separate file): end-to-end balancer convergence with
rack awareness ON and OFF variants. Extract sim fixture to a shared
header so the pinning sim and the existing simulator test share the
same harness.

Topic properties tests: serde roundtrip with/without
replicas_preference, backward compatibility v13 -> v14.
